Mike is right: I get so focused on things that when a two-line change doesn't work and I spend a bunch of time figuring out why, I then waste even more time whinging about it and generally not accepting the brokenness and moving on. My mind is still in "this should be two lines" mode, and I can't let it go.
Tried to fix it so that hovering over transparent portions of a sprite
that is an image doesn't cause it to think it's being moused-over.
There seems to be a bug preventing me from fixing it (as well as bad API
design that caused me to waste time even figuring out that there was a bug)
so I'm going to assume that when that bug is fixed my workaround won't
be necessary. I did restructure a thing or two to be a little cleaner.