Further revamping. Based on the way GWT modules share code and data (as far as I can tell, they don't) we need to create an additional layer of dispatch whereby we use the same module on each of our pages, but that module looks at an additional parameter supplied by the page and dynamically loads up the appropriate "true" entry point from the same codebase used by all of the application.
This may turn out to be a bad idea but I asked the GWT guys so hopefully
they'll enlighten me as to the One True Way (tm) to do what I'm trying to do. I
need to find out more about how JavaScript code is dynamically loaded by GWT
(what they call "deferred binding" and have no documentation for). I want to
make sure that I'm not loading the JavaScript code for the whole nine yards
when someone goes to the inventory page, say, but it may be that all code for a
module has to be loaded all at once.

Anyhow, we now have something that's starting to look like a website. You can
log in, see your inventory and see a placeholder profile page. Next week I'll
get the Flash client jammed into the main page and we'll really be cooking.