Got media uploading all wired together. This involves some deeply twisted business.
First the GWT application creates an anonmyous function and jams it into a
property in the top-level window so that we can call back into GWT from random
Javascript.

Then it creates a <form> element on the fly that is targeted to an invisible
iframe defined in the top-window. The user then clicks a button that submits
that form, causing the iframe to make a POST request to the game server.

A servlet running in the game server receives the file, computes its hash,
groks its mime type, stuffs it into the "media server" (currently a directory)
and then returns HTML that contains Javascript that calls back into the GWT
application via our previously created top-level function to let it know that
the upload is complete and to configure the MediaItem with its media hash and
mime type.

Then the user finishes configuring the item via the GWT application, which
submits a GWT/RPC request to actually create the digital item which now
references the blob of data that was previously uploaded via the POST.

Et voila, a fully operational digital item comes into being. Pillows and
superglue indeed.