A lot of the parlor/table/game stuff uses Names only to identify players. In msoy, display names are not unique, so this presented a problem. Rather than rework all the old stuff, created a MemberName class that encapsulates an immutable memberId. Two MemberNames are equal if they have the same memberId (unless both are -1, in which case the user is a guest and equality will be based on actual usernames.)
This seems to work, although I just thought of an issue: sometimes the
equality of Names matters, sometimes their Comparable-ness. For instance,
if a name is used as a key in a DSet, then the compareTo() method will
be used to find equal keys. Right now, I'm not overridding compareTo()
so that names will continue to display in alphabetical order in lists,
but I may need to make that change... in which case your friends list
would be sorted by memberId rather than by name, which sucks.