Ok, I like this better. When we authenticate, we place the String version of the account name in the _authdata, or null for guests.
When asked to create the starting username, for members we just go ahead
and make the Name version of the account name that will be used, but
for guests we create a guest MemberName.

When resolving, we know that if the account name is a MemberName object,
we are resolving a guest, and we use the same object as the guest's
visible name. Members get their MemberName created during resolution.