No significant changes.
The HTML being generated does not seem to launch msoy in IE.
I can't figure out why, and of course IE has zero facilities to
figure out why. I was loading the same page in firefox so that I could
use FireBug to at least inspect the document and see what was actually
being written.

I fuxed with it, and did not get anywhere, which sucks.