If one belives the InetAddress documentation, since the query to a DNS server comes back with a time to live expressing exactly how many seconds this data is good for, there is no there is no general rule to decide when it is safe to remove cache entries -- and so it caches them forever.
This apparently is a security feature to prevent DNS spoofing and the proper operation of the Internet.

Since Amazon S3 can and does modify their DNS records with relative regularity, we must not cache query results forever. We should do this everywhere at some future date, as we rely on the ability to change DNS records according to the published TTL as needed.