Altered the way submarines move to a way that seems satisfactory.
I've made the decision that the game will really not share much state, but
rather each client participates in a shared hallucination. They'll set up
the board and the initial player positions by convention and then just send
movement updates.

Here's how that works:
- client 0 starts things off by starting a ticker on the server every 100ms.
- every other player simply passes their keyboard events directly to the
  shared gameobject as messages.
- each client receives the messages and updates their local state for
  the appropriate submarine, according to a very simple rule:
  a submarine may only move once per tick.  (Submarines may re-orient as
  often as desired during a tick).

Not yet done: torpedos, and resolving torpedo hits at the end of each tick.

This was probably necessary anyway, as a 100x100 board has 10,000 tiles on
it and there's really no point in making the server care about that big
fat array (which it would if we set it as a property) when clients aren't
going to be able to join an in-progress game anyway.