This seems to work pretty well.
A tick starts, we create a new bucket. All events that arrive during
that tick are put in the bucket. Once the next bucket is started we are
free to render the last bucket.

Note that player actions are still queued by their character, so if you
send 10 left requests that arrive during 1 tick, they'll still play out
during subsequent ticks.