Virtualized portals.
Portals and Furni were extremely similar, except that I couldn't
have them both extend from the same base class because Portals come
from whirled/spot. This resulted in some duplicated code, and every time
I added a new attribute to furni it tended to need to go in portal as well.

I was planning on making the distinction minimal to the user anyway:
they can just take a piece of furniture and say that it's a doorway.
So I realized that I could just rip out all the explicit portal info
and save everything as furniture, and build the portal info needed by
the spot system on the fly out of the furniture data when asked.

This really cleaned up a lot.

Your existing portals should be converted into the equivalent furni.