Ok, the problem with the avatars was in switching to a cloned avatar. I was never trying to load clones properly.
I've gone ahead and renamed (and reordered) some of the methods in
ItemRepository. loadOriginalItem() looks for originals, loadClonedItem()
looks for clones, and loadItem() tries to load an original and falls
back to a clone. They're all named like that: if a method doesn't
have Original|Clone in its name, it tries both.

More to come: item usage isn't tracked for clones.