Complaining time! I just wanted to add something simple so that right clicking on anything in the world will bring up the standard menu, but with an option to examine that item. This will be how you tag something you see in-world, for example.
Except that fucking flash is broken and doesn't return the lowest hit
object below the mouse, and the sprite containing the item media is never
getting asked if it would like to participate in populating the context menu.

Fixing it, I ran into another joyous flash bug (that I already knew about):
transparent pixels count as "hits" even if you ask nicely for them not to.

So: it's in there as best I can. Right-clicking will add the menu item,
but it won't necessarily provide info for the object you thought was under
the mouse pointer.

Yay! Can I spend a little time like this every day? Oh, I do.