While testing some security things, I discovered that it's possible for a client to miss the game starting event because the .swf is still initializing.
Detect an already-running game and cope, but this will need to change
as my game assumes that everyone sees the same events. Probably what
we'll do is require EZGames to explicitely say "I'm ready" instead of
just having the EZGameController (something the user code doesn't have access
to) do it for them.