One line of code, but with so very much behind it. Avatar labels were showing up initially underneath the avatar media, even though I had code to push them above it. I tried adding a listener to wait until the label is validated (and then queue a future function to read the height value due to IDIOCY at Adobe), but that didn't work. Then I tried not using a Label but a simpler TextField, because the only reason I'm trying to push the label up is that the label is a UIComponent and so gets added to the regular child list, while the Loader object that holds the media is not a UIComponent and so needs to be added to the 'rawChildren' list (don't get be started on that). We can specify drawing orders within those lists, but it turns out that the rawChildren will always be painted over the regular children. So with a regular TextField I can just paint it over the avatar media and not have to push it up. But the TextField looked like ass, so I switched back.
At this point I considered giving up. After all, it's just a little bit
of text overlapping. But I'd already tried this much, and if I was just going
to have to revisit this in the future I'd have to try all this over again.

I went back to the Label and tried a few more things before trying to
immediately validate the label after re-checking and setting the width and
position, and that made it _always_ stay down in the wrong place. So it was
the layout fighting against me and when it looked right- THAT was the bug,
the layout somehow allowing me to re-arrange the children without it
noticing. So we turn off layout and manually assign all positions within
our sprites.