Fixed a few bugs, sorted a few things out.
Perspectivization is not fast, nor does it work on large images.
It also doesn't look very good, because I'm not (yet?) doing any
interpolation.

I'm also kinda inventing the algorithm to fit our needs here, but I'm only
doing left/right perspectivization, not ceiling and floor.

What we'll need eventually is going to be more complicated...
We'll probably want the ability to clip when a perspectivized texture
intercepts another wall. We'll want the ability to wallpaper a texture onto
a wall, but we might run into bitmap size limits (2880 x 2880).
We'll see.