Le grande revamp. A few salient points:
1. Media, CSS, etc. are stored in the main document tree and accessed directly
instead of being part of the GWT application's "public" resources.

2. We no longer use the embedded Tomcat server in the GWT app which turns out
to have been totally pointless because we were just redirecting servlet
requests to the local MSOY server anyway. Now we just talk directly to the MSOY
server from the shell.

As a result of this, you have to run "ant gclient" once before the shell will
work because it has to copy the index.html files into the right place to be
accessible to the shell. I think right now changes to the page specific CSS
files would require running ant gclient which is totally unacceptable (that
takes 5 minutes now instead of a minute twenty) so I'll be moving everything
except the (non-changing) index.html file out of the public directory of each
GWT application.

3. Each "major feature" is its own GWT app and resides under a top-level URL:

/world/index.html
/profile/index.html
/group/index.html
/mail/index.html
/inventory/index.html
/catalog/index.html

There may be some niggling details to be sorted out here (groups and profiles
seem to have completely stopped working), but I'll do that once this is checked
in.

Code that is shared between major features needs to be factored into a separate
package that can be included by both. I have done this with client/item and
need to factor some mail stuff out so that it can be shared by the profile,
group and mail packages.

That's probably it. More later if I think of anything.