Include server.properties so that we can specify the proper port to GShell, modified the gshell target such that one must now call it with -Dpage=PAGE to indicate which page they want to see.
Navigating between pages in the GShell currently annoyingly does not preserve
credentials, I tried figuring out why this broke, but it's not clear (it has
something to do with not using the internal Tomcat server I think, or each
module being under a separate path except that I "fixed" that correctly for the
real client so I don't know why it remains broken for the shell).

In any case, the shell tends to run out of memory if you try to load more than
one page anyway, so it tends to be easier and faster to just quit and restart
it with a new page when you want to work on a different module.

Perhaps we could play around with running everything as one big app in the
shell and separately in the real client.