Eradicated all instances of unsafety. GWT service servlet requests are processed on a servlet thread. These methods can thus only safely do two things:
1. Call directly into a repository to load data.
2. Investigate and manipulate the arguments they were provided.

If they want to do anything that involves the server runtime (calling a method
on a Manager, looking up a MemberObject, etc.) they have to first get
themselves to the distributed object thread.

At some point I'd like to modify GWT to allow us to implement the Async
interface on the server and allow us to annotate or otherwise mark methods that
should be rerouted to the distributed object thread before being invoked so
that we can avoid all of this boierplate ServletWaiter and postRunnable() code.
However, until that time, we should do things safely and properly so that if we
do end up having to release with the current system, it doesn't fall apart as
soon as we have enough usage to cause concurrency problems.