Revamped our recenter() handling. Modified our standard popup to start off screen so that it can compute its size and center itself properly without jerking into position. I'd do that set setPopupPosition() except GWT helpfully protects us from danger by forcing the values provided to it to be >= 0.
Also now that our media and images are all constrained to a known size, we can
avoid the whole deferred recentering entirely. I may also nix the recenter on
arrival of details because that currently does not make any major changes to
the UI and it'd be nice for that not to jitter things the little that it does.