A bit of a revamp to support page-specific contexts which will reference a page-specific translation dictionary and later I'll potentially move more remote services into the page-specific contexts if they're not used globally.
To make this happen, I had to refactor the shared bits out of
mail/profile/group so now we have a msgs package which contains things shared
by those packages (the ability to send invitations basically).

The way this has shaped up, we will have a strict inheritance like so:

shell - msgs - mail
             - group
             - profile
      - item - catalog
             - inventory
      - world
      - game

To avoid the morass of faked multiple inheritance, we'll need to avoid
dependence of catalog/inventory on msgs and mail/group/profile on item. Things
that are truly global can go into shell or util.