Started implementing proper authentication. Also started refactoring various things that were pointlessly starting on the servlet thread, moving to the dobj thread, then immediately to the invoker thread to load data from the database, passing back through two threads when they should simply have accessed the database directly on the servlet thread.
Unless something will play a part in the virtual world "runtime", there is no
need to take a trip through the dobj and invoker threads. We're only now in a
position to better understand what will be all web and what will interplay with
the virtual world, but it looks like just about everything except inventory
management is web-only.

I'll be going through the remainder of our services adding proper
authentication and eliminating interoperation with the dobj thread where
appropriate and we can consider for future features whether they are web-only
or require hanlding by a manager which can work in the dobj world.