Handle keyboard input in TruckOrTreat.as, and send messages to all players about what keys are being pressed by which player. This fixes the problem I was having last week where only the player moving saw his character move. Movement feels and looks rather clunky. It would be nice to remedy that eventually.
One problem I have with only listening to KEY_DOWN KeyboardEvents now is that my
in-browser flash plugin does not seem to keep sending events when the keys are
held down. The standalone flash client does, though. So when playing in the
browser you have to keep tapping the arrow keys if you want to keep moving.