Reverted to my old ways of having the kid class handle its own keyboard input. To notify other players, the sprite's new coordinates are sent rather than directions of movement corresponding to keyboard events. This way seems to result in smoother movement.
Now the board only creates a kid object for the player whose board it is, then
sends data about the created kid to the other players so they can create it with
the same parameters. I need to make a similar change to how respawning works, so
when one player dies both don't go respawning in different random locations!