The return of ScalingMediaContainer. Killed ScalingMediaBox and made a more general MediaWrapper. Use that to get the headshots in the lobby. Annoyance: lobby backgrounds are popping to the front. No idea.
While refactoring the item renderers to use the wrapper, at first they
were so totally broken. The tree renderer was acting strange and slow.
I couldn't even scroll to the bottom of my inventory, because something
would jump it back up. The code between ScalingMediaBox and MediaWrapper
at that point was almost identical except for one thing: the broken one
reported its dimensions as 0 when the media was not yet loaded, the good
one went ahead and reserved the maximum space. That little difference
in the renderer made a huge difference in the Tree, so I had to copy
the functionality into MediaWrapper.