Fixed mousing over transparent pixels. Cleanup.
This fix was mostly already in there, but it would throw security errors.
I've fixed it to not throw security errors, which was one line, but I 
spent considerable time trying to do things "correctly"
(flash["correctly"] === undefined), which is to just make the mouseOver
event not fire over transparent pixels. I had some new ideas to try.

Do not ever try this again. I've wasted more time on this than just about
any other flash problem. I searched and searched the interweb and
could find nobody else that either has this problem or thinks it's a
problem. Amazing!

It's possible that this was a security decision on Adobe's part.
Images can be treated as media or data. You may not cross a security
boundary to get data out of an image. It could be that they were paranoid
about transparent images, because if you could receive mouseOver events on a
component underneath, you're effectively getting data out of the image-
which pixels are transparent and which aren't.

If that's true, that doesn't explain why it doesn't work when there is
no security boundary.