Deflexing. Created my own simple class for moving the actors around.
There was another fun mystery here, which is that I wrote the new class
and tried it out and the sprite that was moving was very choppy and bad
looking. I was perplexed, because the math should have worked out the same.
I did tons of debugging and the new and old ways were updating the
coordinates with the same frequency, the values being set were comprable,
and yet I could watch the two moves side-by-side and one was vastly
better looking.

I finally tracked it down to TimerEvent.updateAfterEvent(), which I don't
think I'd heard of before, but let it be hewn in your minds: if you end
up manipulating the display list during the processing of a TimerEvent,
calling this method on the event will instruct the flash player to
update the display list. LIKE IT SHOULD. Fuck.