- Added a method to explicitly grab control without having to set up   a timer, and a callback when control is assigned.
I guess we don't have a system for losing control yet. We may want to
have the ability for an instance to be able to cede control to another
instance.

It might be nice to add a controlWasLost callback now, even if
we don't really use it, so that user code could do the right thing if
we did in the future decide to pull control away from connected players.