I've got my FDR hat on again and I'm coming at you with a New Deal.
We're nixing the *Context objects that are passed around everywhere and instead
moving to a set of classes with static members which we'll reference directly
as appropriate. *Context classes became C* classes which contain the same
stuff, we just save ourselves the trouble of keeping and passing references
around everywhere.

Plus, since GWT is sneaky and statically compiles the results of any call to
C*.msgs.messageString() methods right into the code, it's less weird that you
could do _ctx.imsgs.someString() and have that work even though imsgs was a
non-static member of _ctx.

In defense of the Context pattern as used elsewhere, it is still very useful
when one is defining a set of reusable services like we do in Narya, Vilya,
etc. but in this case we control the horizontal and we control the vertical so
we can deal with the fact that every single one of our classes now has the
implicit non-obvious precondition that its associated context classes must have
been initialized before it is used.