It turns out that in our content packs we'll need to do some event pass-through similar to what I've been doing in the "Stealer".
Rather than do the strange stuff I was doing to make the pass-through work,
just make the event do a few sneaky things so that the passthrough code
reduces down to listening for the event and redispatching a clone, and
that's it.

I also made the class non-dynamic, because I like my buddy the compiler.