Nix quick tags for now.
No finger pointing indended, but for future reference: if something requires a
round trip to the server to get data, stop and ask yourself "Is this
information absolutely essential? Has the user taken an action that results in
them definitively wanting to see this information right now?" If the answer is
no, then think about whether that data loading could be defered until we know
that the user actually cares.

In the case of the quick tags, we will rewrite the tag entry panel to work like
Google's auto suggest feature such that if the user clicks the cursor into the
tag box, we can *then* issue a request to get their quick tags and display them
below the entry box, pruning the displayed tags as the user types letters. Then
we don't issue a pointless request for this tag info every time the user
inspects an item and we provide a less cluttered and more DWIM user interface
in the process.

Thus ends today's lesson in server resource parsimony and user interface
design.