A few things: - The controlled panel is the control bar, not the stage.   There is a hierarchy to controllers, and the ControlBarController should not   hear about CommandEvents dispatched from places other than the control bar. - The two methods being wrapped in adapters should be protected. They are   not part of the public interface and you are bypassing their normal   visibility by passing pointers to them to the adapters.
Also: all locations should be in the previous stack. Recent scenes should
be modified as well to be recent locations. If I'm in a lobby,
the control bar should say "Reversi Lobby" or something, and clicking back
should take me back to the last room. More importantly, if I'm in a game,
clicking back should take me back to the lobby.
It doesn't look like PlaceObject has a string name field, but I think we
should add it.