Haul in Aperture (http://aperture.sourceforge.net/):
    Aperture is a Java framework for extracting and querying full-text content
    and metadata from various information systems (e.g. file systems, web sites,
    mail boxes) and the file formats (e.g. documents, images) occurring in these
    systems.

In particular, Aperture is capable of cheaply, magically detecting the mime
type of a huge assortment of files based on file(1)-style magic -- which
is exactly what I want to use it for. This is very useful for adroitly
detecting the mime-type of user-uploaded content, as we can't particularly
depend on the user's claims.

In limited testing, Aperture performed particularly well:
   http://fredeaker.blogspot.com/2007/02/magicmimetypeidentifier-update.html

Hopefully the very fancy metadata/data extraction/querying capabilities will
be handy elsewhere in metasoy.

Please don't kill me for this massive array of dependencies.
It's a very nice looking library, I promise.