Oh the hackery. Between Macromedia's broken Flash plugin, web brower incompatbilities, GWT engineer's affection for private variables and a healthy helping of bad design decisions made over the glorious history of the web, it's a miracle anything gets done.
We can see all popups over Flash and Java on Linux now, but conveniently one
cannot see the text entry cursor. So it is not immediately apparent that one
can type in text fields that have focus in that situation. I suppose I should
be thankful that this works at all.

By the way, nothing works on IE at the moment. Yay! God knows whether it works
on Safari or Opera.