News flash: parentChanged() does not behave like Java's addNotify()/removeNotify(), if you're still attached to your parent you will not be notified when your parent is detached. There's a new ADDED_TO_STAGE and REMOVED_FROM_STAGE, that were added AFTER the initial release of the flash 9 player, but newer versions have it. WTF were you supposed to do before that?
I cleaned up a few things that needed it now, but we'll have to remember
to fix the rest up at some point.

Also, the event names match the "legacy" ADDED and REMOVED events (which
are eventy versions of parentChanged() for interested parties other than
yourself), which aren't really legacy since AS3 IS A NEW BLEEPING LANGUAGE.

So, these asshats matched the names, but I don't think the usage is matched
because ADDED is (I believe, I'm not testing now) dispatched before you're
added. The REMOVED_FROM_STAGE is definitely dispatched before you're
removed and should be called WILL_REMOVE_FROM_STAGE or something.

Note further the head-up-the-assness: there are flex events overlayed on
top of these events called ADD and REMOVE, which are dispatched *after*
the corresponding actions. Someone needs to go over to Adobe and teach
them what the different tenses of the English language mean.