Another checkpoint commit for decor items: they're now hooked up to work with scenes and the room editor. Decor items can be selected in the room editor, and attached to rooms as background (but unlike regular furni, they can't be deleted). When the room is saved, decor data gets written out into the scene model, and pushed out to other clients.
And there was much rejoicing. (Yey...)

An SWF-in-GWT editor will need to happen next. And some cleanup: room sizes are duplicated 
in the scene model, and that needs to be removed (but don't want to break backward compability 
with existing legacy background furni); also, model attribute updates need to be validated on 
the server (i.e. is the decor valid for this player?), which is currently not happening at all.