Realized that there could be a problem: Your client might receive the event indicating that it's been assigned control, but trying to dispatch that to the entity might be a no-op if the entity is not yet loaded (for example, if you're loading the decor for the room, or if the entity media is simply not initialized yet).
So now I *always* redispatch gotControl() to any entity when we see on
the client that it already has control as a result of the entity requesting
control explicitly or implicitly. This *will* be a problem for entities
that require control, were compiled against older sources, and react in their
GOT_CONTROL handlers in a way that implicitly requires control. There
are all of like 0 of those out in the whirled right now, and if there were
any they're surely made by mdb or I, so I think I'm safe not wrangling with
the ackwards-bombatibility on this one.