Try running the world at 50 frames per second.
This is for Ian, and his games, so that things look better.

<Rant>
The flash player isn't very good at this. It seems that no matter what the
framerate is always below the specified rate. I seriously found posts a
long time ago where some flash coders "fixed" this problem by simply cranking
up their frame rate to 120 and busy-waiting during the extra cycles to
realize their exact desired framerate.

I'm wary of this change, as already I've seen instances of the browser not
responding to seemingly important things like mouse interactions with the
browser UI when the flash player is hogging CPU. But let's check it out.