performance enhancement for the UD engine.
Flash has to scale and redraw each image that has been added to a sprite that is being scaling and
redrawn - so we're adding one more image to the level swfs: a flattened view of the rough, track 
and wall layers.  This way only one image is being redrawn each frame, instead of three.  I still
need access to the original, separate image so I can detect the edge of the track, and where walls
are correctly.

This brought average processor usage during gameplay from ~28% to 20-22% on the standalone release
player on my machine.  More improvements to follow