Fixed bug that causes sprites to highlight when selecting menus and stuff.
Flex just doesn't work well with security restrictions.
Because we load secure children into the roomview, we cannot depend on
mouse overs to tell us when the mouse is inside us or not, so instead we
poll, checking on every ENTER_FRAME. Since we're not using standard mouse
stuff, we don't get all the niceties, and we have to check whether
there's actually another component above us specifically. PITA.