A bit fat reorganization of our data directories to eliminate some old weirdnesses (item/web) and to get everything that should be in the data package into the data package, with some subpackages to allow for the fact that GWT doesn't support annotations yet. So the NWO is:
foo.data - classes used by Java and ActionScript but not GWT; classes in this
package that are not needed by ActionScript can be annotated
@ActionScript(omit=true)

foo.data.all - classes needed by Java, ActionScript and GWT

foo.data.gwt - classes needed by Java and GWT but not ActionScript (we can't
combine this with the .all directory because we can't annotate the classes not
needed by ActionScript with @ActionScript(omit=true))

The one exception to all of this is the web/data package which is really
"web.data.gwt" but the "web" package is all GWT so it seemed pointless to have
web.data.gwt when that's where everything would be. Plus GWT will some day
support annotations and we'll get rid of all this bullshit and just hava "data"
like God and James Gosling intended.

It occurs to me that we could in fact be sneaky and have everything use the
.data package but still separate the source files into these subdirectories.
The compiler doesn't care what directory a file is in, only GWT and the Java ->
ActionScript processing task care about directory structure. I'll have to think
about whether I like that trickery or whether it would be confusing to look at
the package for something and discover that the file was actually in a
subdirectory because of some obscure technical restriction.