Various and sundry bits: changed MINIMIZATION_CHANGED to MINI_WILL_CHANGE to reflect the fact that our notification comes in before our size actually changes. I even changed the Javascript call to take place after we resize the Flash DOM element, but clearly that doesn't happen immediately, so best to have the API reflect reality.
Also fixed AbstractRoomView.scrollRect handling. It was setting the scrollRect
to something wild and crazy when the scrollRect should be set (in non-screen
coordinates) to the size of the "window" onto the whole view. It now does that,
and it does the right thing when the scene size is smaller than our actual
size, which would obviate the need for the PlaceBox mask except we need that
when we have no scrollRect so we'll keep it.

Also a few other inconsequential twiddlings.