In-world door editor, first pass. Right-click on an existing door, and select "set door target". An info pop-up will appear. Move to a new target room, and hit the big "set and return" button on the pop-up. You're done!
It does basic sanity checking before setting the door target (i.e., is the door furni still 
there, and is it still a door?), but doesn't attempt any kind of locking. Also, this only 
works with existing doors - turning a piece of furni into a door will be coming soon.