Having a WorldGameConfig reference in MemberObject opens a can of dependency worms that is best avoided. Since we only actually reference that field in a single place in the entire code base and that place is in Flash no less, we'll just cast it to a WorldGameConfig there and everyone else can remain blisfully unaware of the whole Parlor framework and vast swathes of code.
This means that if you are *in* a WorldGame and try to use Swiftly or the Admin
dashboard panel, it will fail (because that field will actually hold a
WorldGameConfig instance). If that turns out to be an annoyance we can refactor
the world game stuff to request its config based on the oid when it needs it
rather than jamming it into the MemberObject where everyone has to be prepared
to see it whether they care or not.