Route all info and error feedback through static methods in MsoyUI so that we can change the way that feedback is communicated en masse and without fuss.
Also eliminated AlertPopup. Popping up a dialog that has nothing but a message
and an OK button is a crime against user interface humanity. We will at some
point move to a new hotness feedback system like Growl where feedback messages
trot animatedly onto the screen to announce their presence and then depart in
an equally civilized manner, without ever adding the insult of forcing me to
hunt down and click a button to the injury of whatever error I'm already
struggling to cope with.