Checkpoint commit for server notifications.
Server can notify the client about interesting events by creating new
Notification objects and passing them to MemberObject.notify(). On the client
side, the result will be a custom UI baloon, potentially with buttons or other
widgets, that pops up in the corner of the view. The user can interact with the
baloon, or ignore it, in which case it will disappear after a few seconds. 

Internally, notifications are intended for short-lived messages. They are
stored in a simple DSet on the member object, and never touch the database - 
so if the client disconnects, they will be lost.

The main impetus for this was to display a nice notification when a new friend
accepts our invitation, so we can go visit them right away. But right now, for
testing, the notification is hooked up to the "friend logged on" event as well. 
It's kinda neat, we may want to consider keeping it. (By the way, the "visit" 
button doesn't do the right thing yet. But it will! :)