Don't strip too many caps.
This feature was nice on yohoho, where it stopped some people from
chatting in all caps. Sometimes, someone wants to type in all caps
for emphasis or humor, and it was annoying to have that stripped.
On whirled, there are 10,000 more annoying things you can do, so
allow all-cap utterances.

I'm not sure this is a good idea, and I argued with myself whether
the code should just be moved to the YoChatDirector, but didn't want
to start an email discussion with all three projects as to what
color to paint this bikeshed, so I just did this. If you have an
opinion about this, please email me privately.