Some tidying up. The sad thing is that I spent the last two hours trying to wrangle a way to make the Java game-client.jar not 1.5 megs but the power of Proguard is not quite up to the task of handling having some classes in game-client.jar and then others loaded later from the game itself (ones that were not used in game-client.jar).
If I could get Proguard to only keep classes that were needed as if it were
pruning dead code, but then keep the entire classes without then resolving all
the classes referenced by the not-pruned code, we'd be set. But Proguard won't
do that. If I have it keep the code in those classes, it insists on keeping all
the code referenced by that code. (Which is generally sensible, but we know
what we're doing in this case.)

I was somewhat tempted to move all the Java game client stuff into Whirled and
Proguard everything into the game media but that would add ~300k to every game
and require that we proxy the media from S3 because then the game would have to
be downloaded from the game server's IP.

I may be able to go in with a finer toothed comb and prune parts of the
nenya-media and vilya-parlor libraries that we don't need and get things down
closer to 1M which would be nice.