Make the confirmation dialog about moving items that you're trying to add less wide, so that it fits in the 300px-wide minimized flash client.
Another approach to this would be to center a GWT popup in the browser.  I would have done that, if
I could have easily returned a value from FlashClients.addFurni().  However, we may need to load
the required inventory type on the user's MemberObject, so we can't block in that function in the 
flash side in order to return a valid value.  

If it turns out that we still want to do this in a browser-centered GWT popup, I can put together
a generic mechanism for opening that sort of dialog from flash over the bridge, which probably 
wouldn't be very hard.  This was even easier though, so I did it this way for now.