Use a new control for specifying the facing orientation. Added an optional slider for adjusting the scale.
That is actually somewhat boochy, because all the other controls
are transient frobs one can use to try out the avatar, but the
scaling will be permanent.
Perhaps we separate the functionality- to edit scale there's a
button one clicks in the item detail view that replaces the normal
avatarviewer widget with one that allows scaling and hides all the
transient frobs.

Bah.